Fix PR template and Reviewable from merge commits
We recently added a PR template to the main Servo repo to smooth the
contributing workflow. However, this template gets copied into homu's
merge commit messages, generating a lot of noise in the commit logs.

Use heuristics to strip both the PR template and the Reviewable footer.
First, the main body of the PR template and the Reviewable footer are
removed, as these are always the same (except for if the boxes are
checked) and always at the bottom. The other part of the PR template is
a header comment; some users replace it, while others leave it, and even
though the message says to write below that line, some users write above
it. So, simply remove it and trim any extra newlines. These heuristics
should robustly leave the body untouched if the template is changed.
This leaves a much cleaner commit message in the merge commit.

Tested against an excerpt of recent commit messages from Servo.